added notification, notice in use cases and blogger controller

+++ src/features/blogs/application/use-cases/delete-blog.use-case.ts
@@ -1,13 +1,54 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { BlogsRepository } from '../../infrastructure/blogs.repository';
+import { ForbiddenException, NotFoundException } from '@nestjs/common';
+import { Notice } from '../../../../common/notification/notice';
+import { BlogsQueryRepository } from '../../infrastructure/blogs.query-repository';
+import { HTTP_STATUSES } from '../../../../common/utils';
 
 export class DeleteBlogCommand {
-  constructor(public readonly blogId: string) {}
+  constructor(
+    public readonly userId: string,
+    public readonly blogId: string,
+  ) {}
 }
 @CommandHandler(DeleteBlogCommand)
 export class DeleteBlogUseCase implements ICommandHandler<DeleteBlogCommand> {
-  constructor(private blogsRepository: BlogsRepository) {}
+  constructor(
+    private blogsRepository: BlogsRepository,
+    private blogsQueryRepository: BlogsQueryRepository,
+  ) {}
   async execute(command: DeleteBlogCommand) {
-    return await this.blogsRepository.deleteBlog(command.blogId);
+    const notice = new Notice();
+
+    const { userId, blogId } = command;
+
+    const blog = await this.blogsQueryRepository.getBlogById(blogId);
+
+    if (!blog) {
+      notice.addError('Blog not found', 'blogId', HTTP_STATUSES.NOT_FOUND_404);
+      return notice;
+    }
+
+    const isOwnerBlog = await this.blogsQueryRepository.checkOwnerBlog(
+      userId,
+      blogId,
+    );
+
+    if (!isOwnerBlog) {
+      notice.addError(
+        'Blog not owned by user',
+        'userId',
+        HTTP_STATUSES.FORBIDDEN_403,
+      );
+      return notice;
+    }
+
+    const isDeleted = await this.blogsRepository.deleteBlog(command.blogId);
+
+    if (!isDeleted) {
+      notice.addError('Blog not deleted', null, null);
+    }
+
+    return notice;
   }
 }
